+++ NEWS.md
@@ -1,5 +1,10 @@
 # sffdr (development version)
 
+# sffdr 1.1.2
+
+* Added monotonicity option (running min or PAVA)
+* Better handling of weights in kernel density estimation
+
 # sffdr 1.1.1
 
 * Added a weights argument which allows users to input inverse LD scores and fit to all SNPs.